LPS-19701
After reviewing LPS-6125 and LPS-19384 in conjunction with Ray's advice I have removed reference to portlet-body in:

html/common/themes/portlet.jsp
html/common/themes/portlet_content_wrapper.jspf

and left it in:

html/portal/render_portlet.jsp

Seeing that render_portlet.jsp is the only way it is guaranteed in the call path and that it's implementation is better this felt like the best option.

After testing portlet-body only shows up once in the rendered code and behavior is as it should be.