LPS-58337 Selecting a range with 0 results prevents correct rendering of range facets
Analysis
Some of the other display styles like asset_tags.jsp also stop rendering if termCollectors.isEmpty() is true, but their logic is different and therefore they are not affected. Here we iterate over termCollectors to render only tags that indeed matches the search result, so checking termCollectors' emptyness makes seense.

Solution:
I followed the behaviour of modified.jsp on trunk. This doesn't stop rendering if termCollector.isEmpty() is true, so the chance to choose any range or delete currently selected one is given to the user.